pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.4",
    "httpx>=0.28.1",
    "openai>=2.11.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
]
